remove obsolete disposable utilities and tests: eliminate `DisposableBag`, `AsyncDisposableBag`, and `DeferAsync` in favor of `Defer` and `AsyncDefer`; update core and test projects to reflect refactored disposal patterns

+++ AEL.Core.Tests/Stream/ProgressStreamTests.cs
@@ -51,7 +51,7 @@ public sealed class ProgressStreamTests
 		Assert.Equal(inputStream.Length, bytesReadOverall);
 		return;
 
-		void WriteProgress(int bytesRead)
+		void WriteProgress(long bytesRead)
 		{
 			bytesReadOverall += bytesRead;
 		}
@@ -110,7 +110,7 @@ public sealed class ProgressStreamTests
 		Assert.Equal(inputStream.Length, bytesReadOverall);
 		return;
 
-		void WriteProgress(int bytesRead)
+		void WriteProgress(long bytesRead)
 		{
 			bytesReadOverall = bytesRead;
 		}
@@ -143,7 +143,7 @@ public sealed class ProgressStreamTests
 		Assert.Equal(inputStream.Length, bytesReadOverall);
 		return;
 
-		void WriteProgress(int bytesRead)
+		void WriteProgress(long bytesRead)
 		{
 			bytesReadOverall = bytesRead;
 		}
@@ -206,7 +206,7 @@ public sealed class ProgressStreamTests
 		Assert.Equal(inputStream.Length, bytesReadOverall);
 		return;
 
-		void WriteProgress(int bytesRead)
+		void WriteProgress(long bytesRead)
 		{
 			bytesReadOverall += bytesRead;
 		}
@@ -239,7 +239,7 @@ public sealed class ProgressStreamTests
 		Assert.Equal(inputStream.Length, bytesReadOverall);
 		return;
 
-		void WriteProgress(int bytesRead)
+		void WriteProgress(long bytesRead)
 		{
 			bytesReadOverall += bytesRead;
 		}
